Moved breakdown store to common directory
