+++ src/store/breakdown/costOverview/awsCostOverview/awsCostOverviewWidgets.ts
@@ -1,48 +1,48 @@
 import { tagPrefix } from 'api/queries/query';
 import { ReportPathsType, ReportType } from 'api/reports/report';
-import { CostOverviewWidgetType } from 'store/costOverview/common/costOverviewCommon';
+import { CostOverviewWidgetType } from 'store/breakdown/costOverview/common/costOverviewCommon';
 
-import { GcpCostOverviewWidget } from './gcpCostOverviewCommon';
+import { AwsCostOverviewWidget } from './awsCostOverviewCommon';
 
 let currrentId = 0;
 const getId = () => currrentId++;
 
-export const costWidget: GcpCostOverviewWidget = {
+export const costWidget: AwsCostOverviewWidget = {
   id: getId(),
   reportPathsType: ReportPathsType.ocp,
   reportType: ReportType.cost,
   type: CostOverviewWidgetType.cost,
 };
 
-export const accountSummaryWidget: GcpCostOverviewWidget = {
+export const accountSummaryWidget: AwsCostOverviewWidget = {
   id: getId(),
   reportSummary: {
     reportGroupBy: 'account',
     showWidgetOnGroupBy: ['region', 'service', tagPrefix],
   },
-  reportPathsType: ReportPathsType.gcp,
+  reportPathsType: ReportPathsType.aws,
   reportType: ReportType.cost,
   type: CostOverviewWidgetType.reportSummary,
 };
 
-export const regionSummaryWidget: GcpCostOverviewWidget = {
+export const regionSummaryWidget: AwsCostOverviewWidget = {
   id: getId(),
   reportSummary: {
     reportGroupBy: 'region',
     showWidgetOnGroupBy: ['account', 'service', tagPrefix],
   },
-  reportPathsType: ReportPathsType.gcp,
+  reportPathsType: ReportPathsType.aws,
   reportType: ReportType.cost,
   type: CostOverviewWidgetType.reportSummary,
 };
 
-export const serviceSummaryWidget: GcpCostOverviewWidget = {
+export const serviceSummaryWidget: AwsCostOverviewWidget = {
   id: getId(),
   reportSummary: {
     reportGroupBy: 'service',
     showWidgetOnGroupBy: ['region', 'account', tagPrefix],
   },
-  reportPathsType: ReportPathsType.gcp,
+  reportPathsType: ReportPathsType.aws,
   reportType: ReportType.cost,
   type: CostOverviewWidgetType.reportSummary,
 };
