+++ src/store/breakdown/costOverview/azureCostOverview/azureCostOverviewCommon.ts
@@ -1,4 +1,4 @@
-import { CostOverviewWidget } from 'store/costOverview/common/costOverviewCommon';
+import { CostOverviewWidget } from 'store/breakdown/costOverview/common/costOverviewCommon';
 
 export const azureCostOverviewStateKey = 'azureCostOverview';
