+++ src/store/breakdown/costOverview/gcpCostOverview/gcpCostOverviewWidgets.ts
@@ -1,48 +1,48 @@
 import { tagPrefix } from 'api/queries/query';
 import { ReportPathsType, ReportType } from 'api/reports/report';
-import { CostOverviewWidgetType } from 'store/costOverview/common/costOverviewCommon';
+import { CostOverviewWidgetType } from 'store/breakdown/costOverview/common/costOverviewCommon';
 
-import { AwsCostOverviewWidget } from './awsCostOverviewCommon';
+import { GcpCostOverviewWidget } from './gcpCostOverviewCommon';
 
 let currrentId = 0;
 const getId = () => currrentId++;
 
-export const costWidget: AwsCostOverviewWidget = {
+export const costWidget: GcpCostOverviewWidget = {
   id: getId(),
   reportPathsType: ReportPathsType.ocp,
   reportType: ReportType.cost,
   type: CostOverviewWidgetType.cost,
 };
 
-export const accountSummaryWidget: AwsCostOverviewWidget = {
+export const accountSummaryWidget: GcpCostOverviewWidget = {
   id: getId(),
   reportSummary: {
     reportGroupBy: 'account',
     showWidgetOnGroupBy: ['region', 'service', tagPrefix],
   },
-  reportPathsType: ReportPathsType.aws,
+  reportPathsType: ReportPathsType.gcp,
   reportType: ReportType.cost,
   type: CostOverviewWidgetType.reportSummary,
 };
 
-export const regionSummaryWidget: AwsCostOverviewWidget = {
+export const regionSummaryWidget: GcpCostOverviewWidget = {
   id: getId(),
   reportSummary: {
     reportGroupBy: 'region',
     showWidgetOnGroupBy: ['account', 'service', tagPrefix],
   },
-  reportPathsType: ReportPathsType.aws,
+  reportPathsType: ReportPathsType.gcp,
   reportType: ReportType.cost,
   type: CostOverviewWidgetType.reportSummary,
 };
 
-export const serviceSummaryWidget: AwsCostOverviewWidget = {
+export const serviceSummaryWidget: GcpCostOverviewWidget = {
   id: getId(),
   reportSummary: {
     reportGroupBy: 'service',
     showWidgetOnGroupBy: ['region', 'account', tagPrefix],
   },
-  reportPathsType: ReportPathsType.aws,
+  reportPathsType: ReportPathsType.gcp,
   reportType: ReportType.cost,
   type: CostOverviewWidgetType.reportSummary,
 };
