+++ src/store/breakdown/historicalData/awsHistoricalData/awsHistoricalDataCommon.ts
@@ -1,4 +1,4 @@
-import { HistoricalDataWidget } from 'store/historicalData/common/historicalDataCommon';
+import { HistoricalDataWidget } from 'store/breakdown/historicalData/common/historicalDataCommon';
 
 export const awsHistoricalDataStateKey = 'awsHistoricalData';
