+++ src/store/breakdown/historicalData/azureHistoricalData/azureHistoricalDataWidgets.ts
@@ -1,5 +1,5 @@
 import { ReportPathsType, ReportType } from 'api/reports/report';
-import { HistoricalDataWidgetType } from 'store/historicalData/common/historicalDataCommon';
+import { HistoricalDataWidgetType } from 'store/breakdown/historicalData/common/historicalDataCommon';
 
 import { AzureHistoricalDataWidget } from './azureHistoricalDataCommon';
