+++ src/store/breakdown/historicalData/gcpHistoricalData/gcpHistoricalDataCommon.ts
@@ -1,4 +1,4 @@
-import { HistoricalDataWidget } from 'store/historicalData/common/historicalDataCommon';
+import { HistoricalDataWidget } from 'store/breakdown/historicalData/common/historicalDataCommon';
 
 export const gcpHistoricalDataStateKey = 'gcpHistoricalData';
